zfs: Do full clones better
The zfs send command has an option '-I' that will send all of the
snapshots between two points.  It also has an option '-R' that will send
all of the snapshots, however it also does two things that are
undesirable: it is recursive, and it sends properties.

The other way to send a full image is to just leave out both options and
give a snapshot name.  Unfortunately, this only sends a single snapshot
image.

Work around this by first doing a full send of the first snapshot on the
source volume.  After this is done, do an incremental clone from the
first snapshot to the last, which will include all of the intermediate
snapshots.