Add 'prune' command
The prune command will remove old snapshots for a given prefix.  It
tries to keep older backups, with a frequency decreasing the futher back
we go.  This is done with a hanoi-type algorithm, where we basically
keep the most recent backup of a given population count.

The prune command normally does a dry run, and needs `--really` to
actually remove the snapshots.