Make a `checked_run` command
The `CheckedExt` trait adds a `checked_run` method to `Command` which
cleans up much of the error checking.  There are numerous places in the
code that have the same pattern of running a command, and then checking
the status.  Encapsulate that in this method, which will unify the error
handling.