borg: Bind mount to better location
Instead of bind mounting everything to "/mnt/root" for the borg backup,
choose a name appropriate for the volume being mounted.  This makes for
a small hardcoded list, but prevents any possible conflicts from having
the backups mounted in the same place.

This change will require "home" to be rechunked the next time it is
backed up.