Make convention values optional
The go implementation just sets the values to zero if they aren't
present.  serde_yaml is more strict (fitting with Rust), and they have
to be explicitly marked as optional.  Wrap the values, so any of them
can be left out.  It doesn't catch the case where none of the values are
set, but this can be detected at the end.