clone: Don't clone bookmarks
Now that prune commands create bookmarks for deleted snapshots, skip
these when performing clones.  This doesn't fix bookmarks when making
clones, e.g. if a bookmark is determined to be the latest clone the
clone will fail.  But, it allows normal operation to continue, as rack
won't try to clone all of the bookmarks missing on the destination.