Add Cargo.lock header
Recent versions of cargo add a header to the Cargo.lock file stating
that it is auto-generated.  Now that we've seen a few releases with
this, it is probably the case that more people will run into spurious
updates because of being on the newer version than on the old.  Go ahead
and check in the file with the updated header.